+++ src/components/Form/CardNumberForm.tsx
@@ -2,7 +2,7 @@ import Input from "./Input";
 import FormElement from "../common/FormElement";
 import { FormProps } from "./Form";
 import useCardNumberForm from "../../hooks/useCardNumberForm";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 
 export interface CardFormProps extends FormProps {
   labelContent: string;
@@ -23,32 +23,8 @@ const CardNumberForm = ({
   onValidation,
   onFocus,
 }: CardFormProps) => {
-  const [allInputValid, setAllInputValid] = useState(false);
-  const [inputValidities, setInputValidities] = useState({
-    0: false,
-    1: false,
-    2: false,
-    3: false,
-  });
-  const [errorMessage, setErrorMessage] = useState("");
-
-  const isValidCarNumber = (value: string) => /^[0-9]{4}$/.test(value);
-
-  // NOTE: 각 입력 필드의 유효성 검사 결과를 업데이트하는 함수
-  const updateInputValidity = (index: number, isValid: boolean) => {
-    setInputValidities((prevValidities) => ({
-      ...prevValidities,
-      [index]: isValid,
-    }));
-  };
-
-  useEffect(() => {
-    const allValid = Object.values(inputValidities).every((isValid) => isValid);
-    setAllInputValid(allValid);
-    setErrorMessage(allValid ? "" : "4자리의 숫자를 입력해주세요.");
-
-    if (onValidation) onValidation(allValid);
-  }, [inputValidities]);
+  const { updateInputValidity, isValidCarNumber, setErrorMessage, errorMessage } =
+    useCardNumberForm(onValidation!);
 
   const [isFocused, setIsFocused] = useState(false);
